requirements.txt
Django==1.9.12
django-dajax==0.9.2
django-dajaxice==0.7
django-mongodb-engine==0.6.0
django-rq==0.9.6
djangorestframework==3.6.4
djangotoolbox==1.8.0
Jinja2==2.9.6
jsonschema==2.6.0
mongoengine==0.15.0
pymongo==3.6.1
requests==2.18.4
webencodings==0.5.1
